moved all the logic in a custom hook
